% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gather_data.R
\name{gather_data}
\alias{gather_data}
\title{Gather data needed for project}
\usage{
gather_data(tickers = NULL)
}
\arguments{
\item{tickers}{character vector of tickers to gather. Default is NULL,
  meaning use all companies.}
}
\value{
A data frame of required data.
}
\description{
Gather data needed for project
}

