% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{Add together two numbers}
\usage{
make_plot(x, ticker)
}
\arguments{
\item{x}{A data frame with required data}

\item{ticker}{A character with the desired ticker}
}
\value{
A ggplot object
}
\description{
Add together two numbers
}

