% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/old_grab_data.R
\name{old_grab_data}
\alias{old_grab_data}
\title{Gather data needed for project}
\usage{
old_grab_data(symbols, years)
}
\arguments{
\item{years}{numeric vector like 1999:2000 of years to gather}

\item{tickers}{character vector of tickers to gather. Default is NULL,
  meaning use all companies.}
}
\value{
A data frame of required data.
}
\description{
Gather data needed for project
}

